# Improve performance of string interpolation