--- a/string.c
+++ b/string.c
@@ -2880,16 +2880,17 @@ rb_str_concat_literals(size_t num, const VALUE *strary)
 {
     VALUE str;
     size_t i, s;
+    long len = 1;
 
     if (UNLIKELY(!num)) return rb_str_new(0, 0);
     if (UNLIKELY(num == 1)) return rb_str_resurrect(strary[0]);
 
-    long len = 1;
     for (i = 0; i < num; ++i) { len += RSTRING_LEN(strary[i]); }
     if (LIKELY(len < MIN_PRE_ALLOC_SIZE)) {
       str = rb_str_resurrect(strary[0]);
       s = 1;
-    } else {
+    }
+    else {
       str = rb_str_buf_new(len);
       s = 0;
     }