--- a/test/ruby/test_string.rb
+++ b/test/ruby/test_string.rb
@@ -513,6 +513,11 @@ CODE
     assert_raise(RuntimeError) { 'foo'.freeze.concat('bar') }
   end
 
+  def test_concat_literals
+    s="."*50; "#{s}x".encoding
+    assert_equal("#{s}x".encoding, Encoding::UTF_8)
+  end
+
   def test_count
     a = S("hello world")
     assert_equal(5, a.count(S("lo")))